moved method to utils, clean up
