+++ legend-engine-xt-nonrelationalStore-mongodb-javaPlatformBinding-pure/src/main/resources/core_nonrelational_mongodb_java_platform_binding/test/pureSetup.txt
@@ -86,7 +86,7 @@ MongoDBConnection test::mongodb::testConnection
 {
     database: userDatabase;
     store: test::mongodb::mydatabase;
-    serverURLs: [localhost:51821];
+    serverURLs: [localhost:49400];
     authentication: # UserPassword {
         username: 'sa';
         password: SystemPropertiesSecret
